Classes and Functions
Functies in classes geplaatst waardoor de code overzichtelijker word.
